Fixed bugs
Fixed a fault where random number would be added with +2, allowed for
incorrect max amount to be generated.

Fixed "Your guess is above or below 0...." that was showing every time
you won the game.

Fixed bug where 0 was allowed as max number, now the method returns
itself and starts over if 0 as amount is entered.